refactor: GNB 라는 네이밍으로 변경
